refactor(semanticscholar): extract sanitizers to separate module
Move sanitize_arxiv_id and sanitize_dblp_id to dedicated sanitizers.py
module to keep extractors.py under 500 LOC limit.
